+++ app/views/welcome/index.html.haml
@@ -3,6 +3,8 @@
     .govuk-grid-column-two-thirds
       %h1 Home Page
       - if user_signed_in?
+        %h2
+          = "Welcome back #{current_user.email}"
         %li
           = link_to('Logout', destroy_user_session_path, method: :delete)
       - else
